requirements.txt
absl-py>=0.12.0
asgiref>=3.3.1
astunparse>=1.6.3
beautifulsoup4>=4.9.3
cached-property>=1.5.2
cachetools>=4.2.1
certifi>=2020.12.5
chardet>=4.0.0
click>=7.1.2
colorama>=0.4.4
cycler>=0.10.0
Django>=3.1.7
django-chartjs>=2.2.1
django-crispy-forms>=1.11.1
flatbuffers>=1.12
gast>=0.3.3
google-auth>=1.28.0
google-auth-oauthlib>=0.4.3
google-pasta>=0.2.0
grpcio>=1.32.0
h5py>=2.10.0
idna>=2.10
importlib-metadata>=3.7.3
inflection>=0.5.1
joblib>=1.0.1
Keras>=2.4.3
Keras-Preprocessing>=1.1.2
kiwisolver>=1.3.1
loguru>=0.5.3
lxml>=4.6.2
Markdown>=3.3.4
matplotlib>=3.3.4
more-itertools>=8.7.0
multitasking>=0.0.9
nsepy>=0.8
numpy>=1.19.5
oauthlib>=3.1.0
opt-einsum>=3.3.0
pandas>=1.2.3
pandas-datareader>=0.9.0
patsy>=0.5.1
Pillow>=8.1.2
plotly>=4.14.3
protobuf>=3.15.6
pyasn1>=0.4.8
pyasn1-modules>=0.2.8
pyparsing>=2.4.7
python-dateutil>=2.8.1
pytz>=2021.1
PyYAML>=5.4.1
Quandl>=3.6.1
requests>=2.25.1
requests-oauthlib>=1.3.0
retrying>=1.3.3
rsa>=4.7.2
scikit-learn>=0.24.1
scipy>=1.6.1
six>=1.15.0
sklearn>=0.0
soupsieve>=2.2.1
sqlparse>=0.4.1
statsmodels>=0.12.2
tensorboard>=2.4.1
tensorboard-plugin-wit>=1.8.0
tensorflow>=2.4.1
tensorflow-estimator>=2.4.0
termcolor>=1.1.0
threadpoolctl>=2.1.0
tiingo>=0.14.0
typing-extensions>=3.7.4.3
urllib3>=1.26.3
Werkzeug>=1.0.1
win32-setctime>=1.0.3
wrapt>=1.12.1
yfinance>=0.1.55
zipp>=3.4.1